backend/pyproject.toml
[tool.poetry]
name = "speckit-backend"
version = "0.1.0"
description = "Backend API for the AI-Native Textbook Platform"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
qdrant-client = "^1.7.0"
openai = "^1.3.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["jwt"], version = "^3.3.0"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.10.1"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn:app.main:app --reload"